Download dumps over HTTPS and also one chunk at a time
