+++ listenbrainz_spark/config.py.sample
@@ -37,6 +37,9 @@ FTP_LISTENS_DIR = '/pub/musicbrainz/listenbrainz/'
 
 MAPPING_NAME_PREFIX = 'msid-mbid-mapping-with-matchable'
 
+MLHD_PLUS_DUMP_URI = "https://data.metabrainz.org/pub/musicbrainz/listenbrainz/mlhd"
+
+
 PG_JDBC_URI = ""
 PG_USER = ""
 PG_PASSWORD = ""
