+++ listenbrainz_spark/ftp/download.py
@@ -5,6 +5,8 @@ import time
 import logging
 from typing import List
 
+import pycurl
+
 from listenbrainz_spark import config
 from listenbrainz_spark.ftp import ListenBrainzFTPDownloader, DumpType, ListensDump
 from listenbrainz_spark.exceptions import DumpNotFoundException
@@ -161,27 +163,6 @@ class ListenbrainzDataDownloader(ListenBrainzFTPDownloader):
         self.download_file_binary(filename, dest)
         logger.info(f"Done. Total time: {time.monotonic() - t0:.2f} sec")
         return dest
-
-    def download_mlhd_plus_dump(self, directory):
-        ftp_cwd = os.path.join(config.FTP_LISTENS_DIR, 'mlhd/')
-        self.connection.cwd(ftp_cwd)
-
-        downloaded_files = []
-
-        all_files = self.list_dir()
-        for filename in all_files:
-            if "complete" not in filename:
-                continue
-
-            t0 = time.monotonic()
-            logger.info(f"Downloading MLHD+ listen file {filename} from FTP...")
-            file_dest = os.path.join(directory, filename)
-            self.download_file_binary(filename, file_dest)
-            logger.info(f"Done. Total time: {time.monotonic() - t0:.2f} sec")
-
-            downloaded_files.append(file_dest)
-
-        return downloaded_files
 
     def get_latest_dump_id(self, dump_type: DumpType):
         if dump_type == DumpType.INCREMENTAL:
