--- listenbrainz_spark/mlhd/__init__.py
@@ -1,1 +1,0 @@
-MLHD_DATA_PATH = '/data/mlhd'
