+++ listenbrainz_spark/query_map.py
@@ -28,6 +28,7 @@ import listenbrainz_spark.similarity.user
 import listenbrainz_spark.postgres
 import listenbrainz_spark.troi.periodic_jams
 import listenbrainz_spark.tags.tags
+import listenbrainz_spark.mlhd.setup
 
 functions = {
     'stats.entity.listeners': listenbrainz_spark.stats.listener.entity.get_listener_stats,
@@ -39,7 +40,7 @@ functions = {
     'import.dump.full_newest': listenbrainz_spark.request_consumer.jobs.import_dump.import_newest_full_dump_handler,
     'import.dump.full_id': listenbrainz_spark.request_consumer.jobs.import_dump.import_full_dump_by_id_handler,
     'import.dump.incremental_newest': listenbrainz_spark.request_consumer.jobs.import_dump.import_newest_incremental_dump_handler,
-    'import.dump.mlhd': listenbrainz_spark.request_consumer.jobs.import_dump.import_mlhd_dump_to_hdfs,
+    'import.dump.mlhd': listenbrainz_spark.mlhd.setup.import_mlhd_dump_to_hdfs,
     'import.dump.incremental_id': listenbrainz_spark.request_consumer.jobs.import_dump.import_incremental_dump_by_id_handler,
     'cf.missing_mb_data': listenbrainz_spark.missing_mb_data.missing_mb_data.main,
     'cf.recommendations.recording.create_dataframes': listenbrainz_spark.recommendations.recording.create_dataframes.main,
